+++ static/components/f-current-datetime.js
@@ -1,11 +1,11 @@
-import { toLocal, updateField } from "./utils.js";
+import { getNow, updateField } from "./utils.js";
 import { Input } from "./f-text.js";
 
 customElements.define(
   "f-current-datetime",
   class extends Input {
     init() {
-      this.value = this.value ? format(new Date(this.value)) : format();
+      this.value = formatDatetime(this.value);
       this.isNew = false;
 
       super.init();
@@ -16,7 +16,7 @@ customElements.define(
       // Update the value before submitting the form
       input.form.addEventListener(
         "submit",
-        () => input.value = format(new Date()),
+        () => input.value = formatDatetime(),
         { signal: abortController.signal },
       );
 
@@ -36,13 +36,17 @@ customElements.define(
 
     update(schema, value) {
       const input = this.querySelector("input");
-      input.value = value ? format(new Date(value)) : format();
+      input.value = formatDatetime(value);
       updateField(this, schema, input);
     }
   },
 );
 
 // Get the value in the format "YYYY-MM-DDTHH:MM"
-function format(date = new Date()) {
-  return toLocal(date).toISOString().slice(0, 16);
+function formatDatetime(value) {
+  if (!value) {
+    value = getNow().toISOString();
+  }
+
+  return value.slice(0, 16);
 }